+++ module/applications/sheets/encounter-sheet.mjs
@@ -73,10 +73,6 @@ export default class COEncounterSheet extends COBaseActorSheet {
     context.enrichedNotesPublic = await foundry.applications.ux.TextEditor.implementation.enrichHTML(this.document.system.details.notes.public, { async: true })
     context.enrichedNotesPrivate = await foundry.applications.ux.TextEditor.implementation.enrichHTML(this.document.system.details.notes.private, { async: true })
 
-    // Gestion des défenses
-    context.partialDef = this.actor.hasEffect("partialDef")
-    context.fullDef = this.actor.hasEffect("fullDef")
-
     // Gestion des richesses
     context.hasWealth = this.#checkHasWealth(context.system.wealth)